pyproject.toml
[tool.commitizen]
# name = "cz_customize"
version = "1.0.0"
version_files = ["pyproject.toml", "source/version/__init__.py"]

[tool.ruff]
line-length = 88
ignore = ["F401"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"

# [pycodestyle]
# max-line-length = 88

[tool.poetry]
name = "vasco"
packages = [{include = "source"}]
version = "1.0.0"
description = "ai assistant"
authors = ["victor <vfranktor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
mycroft-mimic3-tts = "^0.2.4"
mycroft-messagebus-client = "^0.10.1"
pyxdg = "0.27"
psutil = "^5.9.8"
watchdog = "^4.0.0"
# lingua-franca = "^0.4.3"
uvicorn = "^0.27.1"
openai-whisper = { git = "https://github.com/openai/whisper.git" }
fastapi = "0.110.0"
pvporcupine = "3.0.0"
# fann2 = "1.0.7"
padatious = "0.4.8"
pyaudio = "^0.2.14"
fasteners = "^0.19"
tornado = "^6.4"
adapt-parser = "^1.0.0"
langchain = "^0.1.9"
langchain-openai = "^0.0.7"
elevenlabs = "^0.2.27"
speechrecognition = "^3.10.1"
msm = "^0.8.9"
python-dotenv = "^1.0.1"
open-interpreter = { path = "../open-interpreter" }

[tool.poetry.scripts]
listener-client = "source.client.listener.__main__:main"     # VOICE
messagebus = "source.messagebus.service.__main__:main" # MESSAGEBUS
core = "source.core.__main__:main"                     # CORE
cli-client = "source.client.text.__main__:main"        # CLI
audio = "source.audio.__main__:main"                   # AUDIO
web-client = "source.web.client.__main__:main"         # WEB


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
